Create Check if array elements are consecutive
Method 2 (Use visited array) 
The idea is to check for the following two conditions. If the following two conditions are true, then return true. 
1) max – min + 1 = n where max is the maximum element in the array, min is the minimum element in the array and n is the number of elements in the array. 
2) All elements are distinct.
To check if all elements are distinct, we can create a visited[] array of size n. We can map the ith element of input array arr[] to the visited array by using arr[i] – min as the index in visited[].